split out glwaveformrenderer.h and glwaveformwidgetabstract.h
