+++ src/waveform/renderers/waveformrendererabstract.h
@@ -1,13 +1,12 @@
 #pragma once
 
-#include <QDomNode>
-#include <QGLContext>
-#include <QOpenGLFunctions_2_1>
-#include <QPaintEvent>
-#include <QPainter>
+#include <qglobal.h>
 
-#include "skin/skincontext.h"
+QT_FORWARD_DECLARE_CLASS(QDomNode)
+QT_FORWARD_DECLARE_CLASS(QPaintEvent)
+QT_FORWARD_DECLARE_CLASS(QPainter)
 
+class SkinContext;
 class WaveformWidgetRenderer;
 
 class WaveformRendererAbstract {
@@ -46,30 +45,4 @@ class WaveformRendererAbstract {
     double m_scaleFactor;
 
     friend class WaveformWidgetRenderer;
-};
-
-/// GLWaveformRenderer is a WaveformRendererAbstract which directly calls OpenGL functions.
-///
-/// Note that the Qt OpenGL WaveformRendererAbstracts are not GLWaveformRenderers because
-/// they do not call OpenGL functions directly. Instead, they inherit QGLWidget and use the
-/// QPainter API which Qt translates to OpenGL under the hood.
-class GLWaveformRenderer : protected QOpenGLFunctions_2_1 {
-  public:
-    virtual void onInitializeGL() {
-        initializeOpenGLFunctions();
-    }
-
-  protected:
-    // Somehow QGLWidget does not call QGLWidget::initializeGL on macOS, so hack around that
-    // by calling this in `draw` when the QGLContext has been made current.
-    // TODO: remove this when upgrading to QOpenGLWidget
-    void maybeInitializeGL() {
-        if (QGLContext::currentContext() != m_pLastContext) {
-            onInitializeGL();
-            m_pLastContext = QGLContext::currentContext();
-        }
-    }
-
-  private:
-    const QGLContext* m_pLastContext;
 };
